backend/requirements.txt
# Core AI & Orchestration
anthropic>=0.39.0
openai>=1.54.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
websockets>=13.1

# Trading & Market Data
alpaca-py>=0.30.1
yfinance>=0.2.48
pandas>=2.2.3
numpy>=2.1.3

# Technical Analysis
TA-Lib>=0.4.32

# Backtesting
backtrader>=1.9.78.123

# Social Media & Web Scraping
praw>=7.8.1
tweepy>=4.14.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
playwright>=1.48.0
requests>=2.32.3

# Sentiment Analysis
textblob>=0.18.0.post0
vaderSentiment>=3.3.2

# Data Processing
python-dotenv>=1.0.1
pydantic[email]>=2.9.2
pydantic-settings>=2.6.0

# Database
sqlalchemy>=2.0.36
supabase>=2.10.0

# Utilities
python-dateutil>=2.9.0
pytz>=2024.2
aiohttp>=3.11.2
apscheduler>=3.10.4

# Code Execution (Phase 2-4)
docker>=7.1.0

# Visualization
matplotlib>=3.9.2
plotly>=5.24.1

# Tracing & Observability (Arize Phoenix)
arize-phoenix>=12.0.0
openinference-instrumentation-anthropic>=0.1.0
opentelemetry-sdk>=1.0.0
opentelemetry-exporter-otlp>=1.0.0
